requirements.txt
requests
pixivpy3
Pillow
